requirements.txt
discord.py
transformers
sentence-transformers
